==> Kolkata/03B-Forecast-PM2.5_ML-Data.ipynb <==
import pickle
import pandas as pd

from glob import glob

path = "./ml_models"
model_directories = glob(path + "/*/")

table_dict = {}

for idx in range(len(model_directories)):
    directory = model_directories[idx]
    model_name = directory.split("/")[-2]
    station_pkls = glob(path + "/" + model_name + "/*_pm25.pkl")
    
    model_data = {}

    for inner_idx in range(len(station_pkls)):
        station_data = pickle.load(open(station_pkls[inner_idx], "rb"))
        model_data[station_data["station_id"]] = station_data["score"]
        
    table_dict[model_name + " Regression"] = model_data

reform = {(outerKey, innerKey): values for outerKey, innerDict in table_dict.items() for innerKey, values in innerDict.items()}

result_df = pd.DataFrame(reform).transpose()

result_df

predictions_df = pd.read_pickle(path + "/WB013_predictions.pkl")

predictions_df
