==> Kolkata/06A-Model-PM2.5-AOD.ipynb <==
import pandas as pd
import numpy as np

import pickle
import glob

import warnings
warnings.filterwarnings('ignore')

import configparser

config = configparser.ConfigParser()
config.read("./config.ini")

mcd19a2_obj = pickle.load(open("./mcd19a2.pkl", "rb"))
mcd19a2_longitude, mcd19a2_latitude = mcd19a2_obj['longitude'], mcd19a2_obj['latitude']

stations_df = pd.read_pickle('../2015-2020-pm25/india_stations.pkl')
stations_df = stations_df[stations_df['City'] == 'Kolkata']
stations_df

station_ids = [file.split("/")[-1][:5] for file in glob.glob("./dl_models/*.pkl")]

np.random.seed(46)
train_stations = list(np.random.choice(station_ids, size=6, replace=False))
test_stations  = list(set(station_ids) - set(train_stations))

train_stations

test_stations

LOCATIONS = {}

def add_in_dict(row):
    LOCATIONS[row[1]] = (row[0], row[2], row[3])

[add_in_dict(row) for row in stations_df[['StationName', 'StationId', 'Latitude', 'Longitude']].values];

def get_nearest_point_idx(latitude, longitude, user_lat, user_lon):
        
    R = 6371000
    lat1 = np.radians(user_lat)
    lat2 = np.radians(latitude)
    delta_lat = np.radians(latitude-user_lat)
    delta_lon = np.radians(longitude-user_lon)
    a = (np.sin(delta_lat/2))*(np.sin(delta_lat/2))+(np.cos(lat1))*(np.cos(lat2))*(np.sin(delta_lon/2))*(np.sin(delta_lon/2))
    c = 2*np.arctan2(np.sqrt(a),np.sqrt(1-a))
    d = R*c
    
    x, y = np.unravel_index(d.argmin(),d.shape)
    
    return x, y

def get_nearest_3x3_grid(data, x, y):
    
    if x < 1:
        x += 1
    if x > data.shape[0]-2:
        x -= 2
    if y < 1:
        y += 1
    if y > data.shape[1]-2:
        y -= 2  
    
    three_by_three = data[x-1:x+2,y-1:y+2]
    three_by_three = three_by_three.astype(float)
    
    not_nans = np.count_nonzero(~np.isnan(three_by_three))
    
    if not_nans == 0:
        return {
            "x": x,
            "y": y,
        }
    else:
        three_by_three_average = np.nanmean(three_by_three)
        three_by_three_std = np.nanstd(three_by_three)
        three_by_three_median = np.nanmedian(three_by_three)
        
        return {
            "x": x,
            "y": y,
            "data": three_by_three,
            "average": three_by_three_average,
            "std": three_by_three_std,
            "median": three_by_three_median
        }    

def nearest_lat_lon(station_lat, station_lon):
    x, y = get_nearest_point_idx(mcd19a2_latitude, mcd19a2_longitude, station_lat, station_lon)
    nearest_lon, nearest_lat = np.round(mcd19a2_longitude[x,y], 8), np.round(mcd19a2_latitude[x,y], 8)
    
    return x, y, nearest_lat, nearest_lon

import datetime

def extract_date_from_file_name(FILE_NAME):
    return datetime.datetime.strptime(FILE_NAME.split('/')[-1].split('.')[1:][0][1:], "%Y%j")

PATH = config['convert']['hdf_path'] + "*.hdf"
FILE_LIST = glob.glob(PATH)

from pyhdf import SD
import multiprocessing as mp

manager = mp.Manager()
rows = manager.list()


def perform_task(idx):

    # ------------- Data Loading Start ------------------

    SDS_NAME = "Optical_Depth_047"
    FILE_NAME = FILE_LIST[idx]
    hdf = SD.SD(FILE_NAME)
    sds = hdf.select(SDS_NAME)

    NAME = FILE_NAME.split('/')[-1]

    data = sds.get()

    attributes = sds.attributes()
    scale_factor = attributes['scale_factor']
    fv = attributes['_FillValue']

    data = data.astype(float)
    data[data == fv] = np.nan
    data = np.nanmean(data, axis=0)

    scaled_data = data * scale_factor

    # ------------- Data Loading End ------------------

    date = extract_date_from_file_name(FILE_NAME)
    
    # ------------- AOD Extraction Start --------------

    for sub_idx in range(len(station_ids)):

        station_id = station_ids[sub_idx]
        station_name, station_lat, station_lon = LOCATIONS[station_id]
        x_coord, y_coord, nearest_lat, nearest_lon = nearest_lat_lon(station_lat, station_lon)

        try:
            fix_station_aod = get_nearest_3x3_grid(scaled_data, x_coord, y_coord)['average'].round(3)

            row = [date.strftime("%Y-%m-%d"), station_id, fix_station_aod, nearest_lat, nearest_lon]
            rows.append(row)
        except:
            pass
        
    # ------------- AOD Extraction End --------------

pool = mp.Pool(mp.cpu_count())
pool.map(perform_task, [idx for idx in range(len(FILE_LIST))])
pool.close()

rows = np.array(list(rows))

aod_df = pd.DataFrame(rows, columns=['date', 'station_id', 'aod', 'latitude', 'longitude'])
aod_df['date'] = pd.to_datetime(aod_df['date'], format='%Y-%m-%d')
aod_df = aod_df.set_index(['date', 'station_id'])
aod_df = aod_df.sort_index()

aod_df

stations_df_pm25_list = pd.read_pickle('../2015-2020-pm25/india_stations_pm25.pkl')
pm25_values_df_list = []

for df in stations_df_pm25_list:
    if (df['StationId'].iloc[0] in station_ids):
        pm25_values_df_list.append(df)

START = config['convert']['start_date']
END = config['convert']['end_date']

def previous_window_impute(station_df, window_length=14):
    column = "PM2.5"
    try:
        for idx, value in enumerate(station_df[column].values):
            if np.isnan(value):
                station_df[column][idx] = station_df[column][idx - window_length]
    except:
        pass
    return station_df

for i in range(len(pm25_values_df_list)):
    df = pm25_values_df_list[i]
    nan_count = df.isna().sum()['PM2.5']
    
    if (nan_count):
        df = previous_window_impute(df)
        nan_count = df.isna().sum()['PM2.5']
        if (nan_count):
            print(df['StationId'].values[0])
    
    pm25_values_df_list[i] = pm25_values_df_list[i][START:END]

pm25_df = pd.DataFrame(columns=pm25_values_df_list[0].columns)

for i in range(len(pm25_values_df_list)):
    pm25_df = pm25_df.append(pm25_values_df_list[i])

pm25_df = pm25_df.reset_index()
pm25_df.columns = ["date", "station_id", "PM2.5"] 
pm25_df = pm25_df.set_index(['date', 'station_id'])
pm25_df = pm25_df.sort_index()

merged_df = pd.concat([aod_df, pm25_df], axis=1).dropna().astype("float64").reset_index()

import random

random.seed(42)
random_state = 42

ratio = 0.2

train_df = merged_df[~merged_df['station_id'].isin(test_stations)]
test_df  = merged_df[merged_df['station_id'].isin(test_stations)]

X_columns = ['aod']
y_columns = ['PM2.5']

X_train, y_train = train_df[X_columns].values, train_df[y_columns].values.reshape(-1,1)
X_test, y_test = test_df[X_columns].values, test_df[y_columns].values.reshape(-1,1)

from sklearn.preprocessing import PolynomialFeatures
from sklearn.linear_model import LinearRegression
from sklearn.tree import DecisionTreeRegressor
from sklearn.ensemble import RandomForestRegressor
from sklearn.svm import SVR
from sklearn.metrics import mean_absolute_error, mean_squared_error, mean_squared_log_error, median_absolute_error
from sklearn.metrics import r2_score, mean_absolute_percentage_error
from sklearn.preprocessing import StandardScaler

def get_polynomial_reg_model():
    
    poly_reg = PolynomialFeatures()
    X_poly = poly_reg.fit_transform(X_train)

    regressor = LinearRegression()
    regressor.fit(X_poly, y_train)

    y_pred = regressor.predict(poly_reg.transform(X_test))
    score = {
        "r2_score": r2_score(y_test, y_pred),
        "MAE": mean_absolute_error(y_test, y_pred), 
        "MSLE": mean_squared_log_error(y_test, y_pred),
        "MdAbsE": median_absolute_error(y_test, y_pred),
        "MAPE": mean_absolute_percentage_error(y_test, y_pred),
        "RMSE": np.sqrt(mean_squared_error(y_test, y_pred)),
        "mean": np.mean(y_test)
    }
    
    return regressor, score, poly_reg

def get_linear_reg_model():
    
    regressor = LinearRegression()
    regressor.fit(X_train, y_train)

    y_pred = regressor.predict(X_test)
    score = {
        "r2_score": r2_score(y_test, y_pred),
        "MAE": mean_absolute_error(y_test, y_pred), 
        "MSLE": mean_squared_log_error(y_test, y_pred),
        "MdAbsE": median_absolute_error(y_test, y_pred),
        "MAPE": mean_absolute_percentage_error(y_test, y_pred),
        "RMSE": np.sqrt(mean_squared_error(y_test, y_pred)),
        "mean": np.mean(y_test)
    }
    
    return regressor, score, None

def get_decision_tree_reg_model():
    
    regressor = DecisionTreeRegressor(random_state=random_state)
    regressor.fit(X_train, y_train)
    
    y_pred = regressor.predict(X_test)
    score = {
        "r2_score": r2_score(y_test, y_pred),
        "MAE": mean_absolute_error(y_test, y_pred), 
        "MSLE": mean_squared_log_error(y_test, y_pred),
        "MdAbsE": median_absolute_error(y_test, y_pred),
        "MAPE": mean_absolute_percentage_error(y_test, y_pred),
        "RMSE": np.sqrt(mean_squared_error(y_test, y_pred)),
        "mean": np.mean(y_test)
    }
    
    return regressor, score, None

def get_random_forest_reg_model():
    
    regressor = RandomForestRegressor(random_state=random_state)
    regressor.fit(X_train, y_train)

    y_pred = regressor.predict(X_test)
    score = {
        "r2_score": r2_score(y_test, y_pred),
        "MAE": mean_absolute_error(y_test, y_pred), 
        "MSLE": mean_squared_log_error(y_test, y_pred),
        "MdAbsE": median_absolute_error(y_test, y_pred),
        "MAPE": mean_absolute_percentage_error(y_test, y_pred),
        "RMSE": np.sqrt(mean_squared_error(y_test, y_pred)),
        "mean": np.mean(y_test)
    }
    
    return regressor, score, None

def get_svr_reg_model():
    
    sc_X = StandardScaler()
    sc_y = StandardScaler()
    X_train_sc = sc_X.fit_transform(X_train)
    y_train_sc = sc_y.fit_transform(y_train)
    
    regressor = SVR(kernel='rbf')
    regressor.fit(X_train_sc, y_train_sc)
    
    y_pred = sc_y.inverse_transform(regressor.predict(sc_X.transform(X_test)))
    score = {
        "r2_score": r2_score(y_test, y_pred),
        "MAE": mean_absolute_error(y_test, y_pred), 
        "MSLE": mean_squared_log_error(y_test, y_pred),
        "MdAbsE": median_absolute_error(y_test, y_pred),
        "MAPE": mean_absolute_percentage_error(y_test, y_pred),
        "RMSE": np.sqrt(mean_squared_error(y_test, y_pred)),
        "mean": np.mean(y_test)
    }
    
    return regressor, score, (sc_X, sc_y)

MODEL_LIST = [
    ('Polynomial', get_polynomial_reg_model),
    ('Linear', get_linear_reg_model),
    ('Decision Tree', get_decision_tree_reg_model),
    ('Random Forest', get_random_forest_reg_model),
    ('SVR', get_svr_reg_model),
]

best = {}
best_score = 9e9

rows = []

for name, model_fn in MODEL_LIST:
    regressor, score, scaler = model_fn()
    row = [f"{name} Regression", score['MAE'], score['RMSE'], score['MSLE'], score['MdAbsE']]
    rows.append(row)
    
    if score['RMSE'] < best_score:
        best['name'] = name + " Regression"
        best['regressor'] = regressor
        best['scaler'] = scaler
        best['score'] = score
        
        best_score = score['RMSE']

model_dfs = pd.DataFrame(rows, columns=["model", "MAE", "RMSE", "MSLE", "MdAbsE"])
model_dfs

print(model_dfs.round(3).to_latex())

import pickle

with open(config['convert']["aod_pm25"], "wb") as file:
    pickle.dump(best, file, protocol=4)

y_columns = ['aod']
X_columns = ['PM2.5']

X_train, y_train = train_df[X_columns].values, train_df[y_columns].values.reshape(-1,1)
X_test, y_test = test_df[X_columns].values, test_df[y_columns].values.reshape(-1,1)

best = {}
best_score = 9e9

rows = []

for name, model_fn in MODEL_LIST:
    regressor, score, scaler = model_fn()
    row = [f"{name} Regression", score['MAE'], score['RMSE'], score['MSLE'], score['MdAbsE']]
    rows.append(row)
    
    if score['RMSE'] < best_score:
        best['name'] = name + " Regression"
        best['regressor'] = regressor
        best['scaler'] = scaler
        best['score'] = score
        
        best_score = score['RMSE']

model_dfs = pd.DataFrame(rows, columns=["model", "MAE", "RMSE", "MSLE", "MdAbsE"])
model_dfs

print(model_dfs.round(3).to_latex())

import pickle

with open(config['convert']["pm25_aod"], "wb") as file:
    pickle.dump(best, file, protocol=4)
