==> Kolkata/09D-Model-Final-Data-2.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

import pickle

import warnings
warnings.filterwarnings('ignore')

import math

plt.rc('text', usetex=True)  
plt.rc('font', family='sans-serif')

results = pickle.load(open("./final_results.pkl", "rb"))

results['Deep Learning']['Bi-Directional LSTM'] = results['Deep Learning']['BiDirectional LSTM']
del results['Deep Learning']['BiDirectional LSTM']

results['Deep Learning']['LSTM Auto-Encoder'] = results['Deep Learning']['LSTM AutoEncoder']
del results['Deep Learning']['LSTM AutoEncoder']

from sklearn.metrics import mean_absolute_error, mean_squared_error, mean_squared_log_error, median_absolute_error, mean_absolute_percentage_error

scores = {}

approaches = list(results.keys())

for approach in approaches:
    scores[approach] = {}

    methods = list(results[approach].keys())
    
    for method in methods:

        y_hat = results[approach][method]['y_pred']
        y_test = results[approach][method]['y_test']

        rmse    = np.sqrt(mean_squared_error(y_test, y_hat))
        mae     = mean_absolute_error(y_test, y_hat)
        msle    = mean_squared_log_error(y_test, y_hat)
        mdabse   = median_absolute_error(y_test, y_hat)
        mape    = mean_absolute_percentage_error(y_test, y_hat)

        scores[approach][method] = {
            "MAE": mae,
            "RMSE": rmse,
            "MSLE": msle,
            "MdAbsE": mdabse,
#             "MAPE": mape
        }

reform = {(outerKey, innerKey): values for outerKey, innerDict in scores.items() for innerKey, values in innerDict.items()}

result_df = pd.DataFrame(reform).transpose().round(3)

result_df

print(result_df.to_latex())
